refactor: 디렉터리명 수정
