+++ client/src/components/RestaurantCandidateList/styles.tsx
